Retrieve information from a conductivity calibration details file

get_cond_details() extracts a specific subset of the information stored in a details file generated by HOBOware during conductivity/salinity calibration.

Usage

get_cond_details(path)

Arguments

path

The path to a details file created by HOBOware during conductivity calibration.

Value

A nested list containing of a small subset of the information in file.

See also

Examples

 file <- system.file(
   "extdata/2023/RB1/2023-06-09/Calibrated/Cond_RB1_2023-06-09_Details.txt",
                     package = "BuzzardsBay")
 do_cal  <- get_cond_details(file)
 cat(yaml::as.yaml(do_cal))
#> cond_calibration:
#>   calibration_points: Start point and End point
#>   start_cal_cond: 40768.0
#>   start_cal_temp: 21.1
#>   start_cal_time: 06/02/23 11:10:00 AM GMT-04:00
#>   end_cal_cond: 39200.0
#>   end_cal_temp: 17.8
#>   end_cal_time: 06/09/23 12:40:00 PM GMT-04:00
#> cond_deployment:
#>   full_series_name: Salinity, ppt
#>   launch_name: BBC3_RB1_20636185
#>   launch_time: 06/01/23 14:41:47  GMT-04:00
#>   logging_interval: 00 Hr 10 Min 00 Sec
#>   logging_interval_min: 10.0
#> cond_device:
#>   product: HOBO U24-002 Conductivity
#>   serial_number: 20636185
#>   version_number: 1.52
#>   header_created: 06/10/19 08:05:37  GMT-04:00