Retrieve information stored a Dissolved Oxygen calibration details file

get_do_details() extracts a specific subset of the information stored in a details file generated by HOBOware during DO calibration.

Usage

get_do_details(path)

Arguments

path

The path to a details file created by HOBOware during calibration.

Value

A nested list containing of a small subset of the information in file.

See also

Examples

 file <- system.file(
   "extdata/2023/RB1/2023-06-09/Calibrated/DO_RB1_2023-06-09_Details.txt",
                     package = "BuzzardsBay")
 do_cal  <- get_do_details(file)
 cat(yaml::as.yaml(do_cal))
#> do_calibration:
#>   start_time: 06/02/23 11:10:00 AM GMT-04:00
#>   start_do_conc: 7.37
#>   start_temperature_c: 22.0
#>   start_salinity_ppt: 28.47
#>   start_meter_titration_value_mg_l: 6.41
#>   start_salinity_correction: 0.8475
#>   end_time: 06/09/23 12:40:00 PM GMT-04:00
#>   end_do_conc: 9.12
#>   end_temperature_c: 18.1
#>   end_meter_titration_value_mg_l: 7.62
#> do_deployment:
#>   full_series_name: DO Adj Conc , mg/L
#>   launch_name: BBC3_RB1_20659182
#>   launch_time: 06/01/23 14:16:43  GMT-04:00
#>   logging_interval: 00 Hr 10 Min 00 Sec
#>   calibration_date: 05/16/23 17:04:40  GMT-04:00
#>   calibration_gain: 1.07718
#>   calibration_offset: -0.02045
#>   logging_interval_min: 10.0
#> do_device:
#>   product: HOBO U26-001 Dissolved Oxygen
#>   serial_number: 20659182
#>   version_number: 1.08
#>   header_created: 03/02/12 14:19:05  GMT-04:00