Lookup paths for data and parameter files

This includes all the paths that relate to a deployment, site, year, or the entire data base.

Usage

lookup_paths(
  base_dir = NULL,
  year = NULL,
  site = NULL,
  deployment_date = NULL,
  deployment_dir = NULL
)

Arguments

base_dir

The base directory for all the data.

year

The four digit year for which paths are to be looked up

site

The site name for which the paths are to be looked up

deployment_date

The deployment date in the form yyyy-mm-dd, where yyyy is a four digit year, mm is a two digit month and dd is a two digit day. Single digit components should be padded e.g. 01 for January.

deployment_dir

(optional) The full path to a deployment's directory. If deployment_dir is used, all other arguments are ignored and instead their values are inferred from this path.

Value

A list of paths and a few related non-path variables. The items returned are all always the same and in the same order but with some inputs the deployment specific items will not be defined so will have NA values. All list items contain a single string.

base_dir

The path to the base directory for the data.

year_dir

The path to a specific year's directory

site_dir

The path to a specific site's directory

md_dir

The path to a metadata / parameter directory. Containing additional information about sites and placements for the given year

deployment_dir

The path to the deployment directory where all the data (raw, calibrated, and QC'd) for a deployment is stored. A deployment is the period a device is continuously in the water preceding data download from the device, typically about a week for this project.

deployment_cal_dir

The deployment calibration directory path

deployment_auto_qc

The deployment Auto QC file path this file is created by qc_deployment() as a permanent record of the output of the automated QC process. It should not be modified.

deployment_prelim_qc

This is the preliminary QC file. It is created by qc_deployment() and is intended to be edited by people during the AC process who will, at completion of that process delete the "preliminary_" from the beginning of the name.

deployment_final_qc

This is the final QC file created from deployment_prelim_qc by people performing the QC.

deployment_metadata

This file contains metadata from the HOBOware Details files and generated by qc_deployment().

deployment_report

An html report with plots to inform the QC process.

preceding_auto_qc

The path to the auto QC file for the preceding deployment if it exists, NA otherwise.

sites

CSV file with sites and associated information. The sites file is maintained by the users.

placements

CSV tracking placements. The placement is the period a specific device is at a site. The placements file is maintained by the users.

import_types

File connecting device models (in placements file) to the import type, an integer, used to import the calibrated data derived from those device models each import type must be supported with a function named import_calibrated_data_{type}. where type the integer. See for example import_calibrated_data_1 which handles the CSV output from HOBOware. The public function import_calibrated_data() handles dispatch to the numbered functions and does some additional formatting and checking of the output.

global_parameters

This YAML (.yml) file sets global parameters for the BuzzardsBay package, most of which affect how flags are calculated. See update_bb_parameters

site_parameters

This YAML takes precedence over global_parameters and allows adjusting them to a particular site

year

NOT a path This is the four digit year in character format

deployment_date

NOT a path. This is the deployment date as a character in yyyy-mm-dd format. Numbers are padded e.g. 01 for January.

site

NOT a path. This is the site name. It is case sensitive and should appear exactly as it does in the sites table (sites).

Details

The easiest way to call it is to specify the deployment directory (deployment_dir) in which case all other arguments are inferred from the path. Otherwise specify at a minimum base_dir, year, and site; and, optionally deployment_date and all paths that can be specified will be.