Locations of NOAA tide stations within Buzzards Bay

Locations of NOAA tide stations within Buzzards Bay

Usage

tide_station_info

Format

tide_station_info

A data frame with 13 rows and 5 columns:

name

Station name. This can be used with rtide::tide_height()

id

Station ID

lat,lon

Station latitude and longitude

predictions

The type of model used to make predictions. One of "Harmonic" or `"Subordinate".

Descriptions from NOAA:

Harmonic stations are stations with tidal harmonic constants and tidal datums. Tide predictions for harmonic stations are generated directly from the harmonic constants. Harmonic stations have the greatest capabilities within the NOAA Tide Predictions service for providing predictions with different data intervals and relative to different tidal datums. These capabilities are controlled within the interface.

Subordinate stations are historic stations which do not have tidal harmonic constants available. Tide predictions for subordinate stations are generated by first generating high/low tide predictions for a designated harmonic station, called the "reference station"; then time and height adjustments are applied to correct the high/low predictions to the tidal conditions at the subordinate station. Subordinate stations only allow for the generation of high/low tide predictions, and heights will be relative to Mean Lower Low Water (MLLW), the standard chart datum for the U.S. coastline.

used

Indicates whether the station is in sites.csv

rtide_name

The name for the station as used by rtide package

Source

<"https://tidesandcurrents.noaa.gov/tide_predictions.html?gid=1403">